pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "botpack"
version = "0.1.0"
description = "Agent asset package manager (skills/commands/agents + MCP)"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "tomli>=2.0.1; python_version < '3.11'",
]

[project.urls]
Repository = "https://github.com/Smarty-Pants-Inc/botpack"

[project.scripts]
botpack = "botpack.cli:main"

[dependency-groups]
dev = [
  "pytest>=9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["botpack"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = ["-q"]
